+++ packages/amplify-category-function/src/provider-utils/awscloudformation/utils/build.ts
@@ -3,8 +3,8 @@ import { buildFunction, BuildRequestMeta } from './buildFunction';
 import { ServiceName } from './constants';
 
 export const buildResource = (context: $TSContext, resource: BuildRequestMeta & { service: string }) => {
-  // no build step for lambda layers
-  if (resource.service === ServiceName.LambdaLayer) {
+  // only build lambda functions
+  if (resource.service !== ServiceName.LambdaFunction) {
     return;
   }
   return buildFunction(context, resource);