+++ packages/amplify-codegen/src/utils/getGraphQLDocPath.js
@@ -1,17 +1,13 @@
-const { join, dirname } = require('path');
+const { join } = require('path');
+const globParent = require('glob-parent');
 
-function getGraphQLDocPath(frontend, schemaLocation) {
-  const graphQLDirectory = dirname(schemaLocation);
-  let subDirectory;
-  switch (frontend) {
-    case 'android':
-      subDirectory = 'com/amazonaws/amplify/generated/graphql';
-      break;
-    default:
-      subDirectory = '';
+function getGraphQLDocPath(frontend, graphQLDirectory, includePathGlob) {
+  if (frontend === 'android') {
+    return join(graphQLDirectory, 'com/amazonaws/amplify/generated/graphql');
   }
-
-  return join(graphQLDirectory, subDirectory);
+  return Array.isArray(includePathGlob) && includePathGlob.length > 0
+    ? globParent(includePathGlob[0])
+    : graphQLDirectory;
 }
 
 module.exports = getGraphQLDocPath;