+++ packages/amplify-provider-awscloudformation/src/utils/admin-helpers.ts
@@ -2,6 +2,7 @@ import { stateManager, $TSContext } from 'amplify-cli-core';
 import aws from 'aws-sdk';
 import _ from 'lodash';
 import fetch from 'node-fetch';
+import { adminLoginFlow } from '../admin-login';
 import { AdminAuthConfig, AwsSdkConfig, CognitoAccessToken, CognitoIdToken } from './auth-types';
 
 export const adminVerifyUrl = (appId: string, envName: string, region: string): string => {
@@ -27,8 +28,11 @@ export async function isAmplifyAdminApp(appId: string): Promise<{ isAdminApp: bo
   return { isAdminApp: !!appState.appId, region: appState.region };
 }
 
-export async function getTempCredsWithAdminTokens(appId: string, print: $TSContext['print']): Promise<AwsSdkConfig> {
-  const authConfig = await getRefreshedTokens(appId, print);
+export async function getTempCredsWithAdminTokens(context: $TSContext, appId: string): Promise<AwsSdkConfig> {
+  if (!doAdminTokensExist(appId)) {
+    await adminLoginFlow(context, appId);
+  }
+  const authConfig = await getRefreshedTokens(context, appId);
   const { idToken, IdentityId, region } = authConfig;
   // use tokens to get creds and assign to config
   const awsConfig = await getAdminCognitoCredentials(idToken, IdentityId, region);
@@ -82,17 +86,22 @@ async function getAdminStsCredentials(idToken: CognitoIdToken, region: string): 
   };
 }
 
-export async function getRefreshedTokens(appId: string, print: $TSContext['print']) {
+async function getRefreshedTokens(context: $TSContext, appId: string) {
   // load token, check expiry, refresh if needed
   const authConfig: AdminAuthConfig = stateManager.getAmplifyAdminConfigEntry(appId);
 
   if (isJwtExpired(authConfig.idToken)) {
-    const refreshedTokens = await refreshJWTs(authConfig, print);
-    // Refresh stored tokens
-    authConfig.accessToken.jwtToken = refreshedTokens.AccessToken;
-    authConfig.idToken.jwtToken = refreshedTokens.IdToken;
-    authConfig.refreshToken.token = refreshedTokens.RefreshToken;
-    stateManager.setAmplifyAdminConfigEntry(appId, authConfig);
+    let refreshedTokens: aws.CognitoIdentityServiceProvider.AuthenticationResultType;
+    try {
+      refreshedTokens = (await refreshJWTs(authConfig)).AuthenticationResult;
+      // Refresh stored tokens
+      authConfig.accessToken.jwtToken = refreshedTokens.AccessToken;
+      authConfig.idToken.jwtToken = refreshedTokens.IdToken;
+      stateManager.setAmplifyAdminConfigEntry(appId, authConfig);
+    } catch {
+      // Refresh failed, fall back to login
+      await adminLoginFlow(context, appId, null, authConfig.region);
+    }
   }
   return authConfig;
 }
@@ -103,21 +112,15 @@ function isJwtExpired(token: CognitoAccessToken | CognitoIdToken) {
   return secSinceEpoch >= expiration - 60;
 }
 
-async function refreshJWTs(authConfig: AdminAuthConfig, print: $TSContext['print']) {
+async function refreshJWTs(authConfig: AdminAuthConfig) {
   const CognitoISP = new aws.CognitoIdentityServiceProvider({ region: authConfig.region });
-  try {
-    const result = await CognitoISP.initiateAuth({
-      AuthFlow: 'REFRESH_TOKEN',
-      AuthParameters: {
-        REFRESH_TOKEN: authConfig.refreshToken.token,
-      },
-      ClientId: authConfig.accessToken.payload.client_id, // App client id from identityPool
-    }).promise();
-    return result.AuthenticationResult;
-  } catch (e) {
-    print.error(`Failed to refresh tokens: ${e.message || 'Unknown error occurred'}`);
-    throw e;
-  }
+  return await CognitoISP.initiateAuth({
+    AuthFlow: 'REFRESH_TOKEN',
+    AuthParameters: {
+      REFRESH_TOKEN: authConfig.refreshToken.token,
+    },
+    ClientId: authConfig.accessToken.payload.client_id, // App client id from identityPool
+  }).promise();
 }
 
 export const adminBackendMap: {