+++ packages/graphql-key-transformer/src/KeyTransformer.ts
@@ -1076,7 +1076,7 @@ function setSyncQueryFilterSnippet() {
     compoundExpression([
       set(ref('filterArgsMap'), ref('ctx.args.filter.get("and")')),
       ifElse(
-        raw(`!$util.isNullOrEmpty($filterArgsMap)`),
+        raw(`!$util.isNullOrEmpty($filterArgsMap) && $util.isNull($ctx.args.lastSync)`),
         compoundExpression([
           set(ref('json'), raw(`$filterArgsMap`)),
           forEach(ref('item'), ref(`json`), [