+++ packages/amplify-category-analytics/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.20.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.20.1...amplify-category-analytics@2.20.2) (2021-02-24)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
+## [2.20.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.20.0...amplify-category-analytics@2.20.1) (2021-02-17)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
 # [2.20.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@2.19.5...amplify-category-analytics@2.20.0) (2021-02-11)