+++ packages/amplify-category-function/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.30.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@2.30.0...amplify-category-function@2.30.1) (2021-02-24)
+
+**Note:** Version bump only for package amplify-category-function
+
+
+
+
+
+# [2.30.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@2.29.0...amplify-category-function@2.30.0) (2021-02-17)
+
+
+### Bug Fixes
+
+* correct function resource name validation ([#6570](https://github.com/aws-amplify/amplify-cli/issues/6570)) ([5737b46](https://github.com/aws-amplify/amplify-cli/commit/5737b46df2ca405cfa0d56e03d5b8e74102cffd4))
+* fix messed up merge ([#6612](https://github.com/aws-amplify/amplify-cli/issues/6612)) ([0e5d85b](https://github.com/aws-amplify/amplify-cli/commit/0e5d85be780c800aad2322ebb2b5598187c97ae8))
+* mock bug fixes and e2e test updates ([#6626](https://github.com/aws-amplify/amplify-cli/issues/6626)) ([af76446](https://github.com/aws-amplify/amplify-cli/commit/af76446d18bf626ca5f91c3ad41081175c959807))
+
+
+### Features
+
+* fully populate mock function environment variables ([#6551](https://github.com/aws-amplify/amplify-cli/issues/6551)) ([dceb13a](https://github.com/aws-amplify/amplify-cli/commit/dceb13a76a85a05940078868a3e2e1ca85656938))
+* Separate prod and dev lambda function builds ([#6494](https://github.com/aws-amplify/amplify-cli/issues/6494)) ([2977c6a](https://github.com/aws-amplify/amplify-cli/commit/2977c6a886b33a38ef46f898a2adc1ffdb6d228b))
+
+
+
+
+
 # [2.29.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@2.28.5...amplify-category-function@2.29.0) (2021-02-11)