+++ packages/amplify-python-function-template-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-python-function-template-provider@1.3.3...amplify-python-function-template-provider@1.3.4) (2021-02-17)
+
+**Note:** Version bump only for package amplify-python-function-template-provider
+
+
+
+
+
 ## [1.3.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-python-function-template-provider@1.3.2...amplify-python-function-template-provider@1.3.3) (2021-02-10)