+++ packages/amplify-util-import/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.12](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.3.11...amplify-util-import@1.3.12) (2021-02-24)
+
+**Note:** Version bump only for package amplify-util-import
+
+
+
+
+
+## [1.3.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.3.10...amplify-util-import@1.3.11) (2021-02-17)
+
+**Note:** Version bump only for package amplify-util-import
+
+
+
+
+
 ## [1.3.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-import@1.3.9...amplify-util-import@1.3.10) (2021-02-11)
 
 **Note:** Version bump only for package amplify-util-import