+++ packages/amplify-util-mock/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.28.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.28.0...amplify-util-mock@3.28.1) (2021-02-24)
+
+
+### Bug Fixes
+
+* handle no mock args ([#6659](https://github.com/aws-amplify/amplify-cli/issues/6659)) ([6b97ffa](https://github.com/aws-amplify/amplify-cli/commit/6b97ffa34d387c6a6e1e494d8f25db33710de5fa))
+
+
+
+
+
+# [3.28.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.27.9...amplify-util-mock@3.28.0) (2021-02-17)
+
+
+### Bug Fixes
+
+* fix messed up merge ([#6612](https://github.com/aws-amplify/amplify-cli/issues/6612)) ([0e5d85b](https://github.com/aws-amplify/amplify-cli/commit/0e5d85be780c800aad2322ebb2b5598187c97ae8))
+* mock bug fixes and e2e test updates ([#6626](https://github.com/aws-amplify/amplify-cli/issues/6626)) ([af76446](https://github.com/aws-amplify/amplify-cli/commit/af76446d18bf626ca5f91c3ad41081175c959807))
+* mock handles and prints child proc errors ([#6601](https://github.com/aws-amplify/amplify-cli/issues/6601)) ([ce075d9](https://github.com/aws-amplify/amplify-cli/commit/ce075d91c0b93885229ab3c6000a450c6b8cc56a))
+
+
+### Features
+
+* fully populate mock function environment variables ([#6551](https://github.com/aws-amplify/amplify-cli/issues/6551)) ([dceb13a](https://github.com/aws-amplify/amplify-cli/commit/dceb13a76a85a05940078868a3e2e1ca85656938))
+* Separate prod and dev lambda function builds ([#6494](https://github.com/aws-amplify/amplify-cli/issues/6494)) ([2977c6a](https://github.com/aws-amplify/amplify-cli/commit/2977c6a886b33a38ef46f898a2adc1ffdb6d228b))
+
+
+
+
+
 ## [3.27.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.27.8...amplify-util-mock@3.27.9) (2021-02-11)