+++ packages/graphql-elasticsearch-transformer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.4](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@4.10.3...graphql-elasticsearch-transformer@4.10.4) (2021-02-24)
+
+**Note:** Version bump only for package graphql-elasticsearch-transformer
+
+
+
+
+
+## [4.10.3](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@4.10.2...graphql-elasticsearch-transformer@4.10.3) (2021-02-17)
+
+**Note:** Version bump only for package graphql-elasticsearch-transformer
+
+
+
+
+
 ## [4.10.2](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@4.10.1...graphql-elasticsearch-transformer@4.10.2) (2021-02-11)
 
 **Note:** Version bump only for package graphql-elasticsearch-transformer