+++ packages/graphql-function-transformer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.8](https://github.com/aws-amplify/amplify-cli/compare/graphql-function-transformer@2.4.7...graphql-function-transformer@2.4.8) (2021-02-24)
+
+**Note:** Version bump only for package graphql-function-transformer
+
+
+
+
+
+## [2.4.7](https://github.com/aws-amplify/amplify-cli/compare/graphql-function-transformer@2.4.6...graphql-function-transformer@2.4.7) (2021-02-17)
+
+**Note:** Version bump only for package graphql-function-transformer
+
+
+
+
+
 ## [2.4.6](https://github.com/aws-amplify/amplify-cli/compare/graphql-function-transformer@2.4.5...graphql-function-transformer@2.4.6) (2021-02-11)
 
 **Note:** Version bump only for package graphql-function-transformer